application: use From trait instead of Into
The reason is that Into<T> does not have a corresponding trait
automatically defined as From<U>. And in turn, the reason for this
is that, besides avoiding a circular reference that would conflict
with existing From impls (and without specialization that'd be an
error), it is not always possible to define a From<U> trait due to
coherence rules, but you can always define an Into<T> trait.

At https://doc.rust-lang.org/std/convert/trait.Into.html you can
read the paragraph below:

"Library authors should not directly implement this trait, but
should prefer implementing the From trait, which offers greater
flexibility and provides an equivalent Into implementation for free,
thanks to a blanket implementation in the standard library."

The greater flexibility comes from the blanket Into implementation.
Additionally, check out the problem about not always being able to
implement From<T> (ie. you can't impl From<T> for Vec<T>) at:

https://doc.rust-lang.org/std/convert/trait.Into.html#implementing-into