misc: add internal nightly feature flag for running benches
Unfortunately benchmarks are not stable and are feature-gated even
with nightly compilers (ie. feature(test) + extern crate test).

With this change, #[bench] and the `Bencher` struct will be hidden
to stable compilers (even when running cargo bench) and only used
with compilers specifying the `nightly` feature (hopefully just
nightly compilers, with a forthcoming addition to the build
process).