extensions: hide internal implementation and make the API more ergonomic
This makes the internal HashMap a hidden implementation detail and
accepts &str rather than just String in methods (with obvious
cloning, but that is something we can consider dealing with later).

The nice thing here is how we can avoid converting a &str into
a String to remove from the HashMap since it can just accept a &str.