http/request: redesign the trait to configure a client with a Request
The previous trait, while working, had some ergonomic issues. For
example, it did require in some cases for the implementation to
create the client, and in others it required a mutable reference or
a move in the generic parameters.

It also used a free-standing function to do this, and in some cases
it required non-obvious types to call it on.

The new trait feels more natural while maintaining the flexibility,
in which we now require a client to exist before configuring our
request to be sent off. This change should convey that impls will no
longer create the clients, but just set them up with whatever is
required to perform a call.