timestamp: handle as many timestamps as possible from UNIX_EPOCH
This includes timestamps before the Epoch in most platforms, at the
price of about half the billions of years in the future by using an
i64 to make conversions.

This uses TryFrom and is unlikely to be used in normal cases, but
could be useful for backfilling reports with negative timestamps in
cases (usually very specialised or in emergency) where Apisonator
would be changed to accept them.