misc: switch usages of HashMap to BTreeMap to avoid pulling in extras
Since most cases where these are used the actual structure is not
important, we can switch them to BTreeMap easily while we think whether
other options (ie. Vec) are better suited to them, and this way we also
ensure we don't need to pull in another dependency like hashbrown to
support it in no_std mode.