misc: remove #[repr(transparent)] instances
They prevent some optimizations and the repr should only really be used
when some FFI or similar layout guarantee is directly required, such as
when passing over typed units. More information about this in urlo at:
https://users.rust-lang.org/t/repr-transparent-why/67636

If we find issues with these we should be restoring them on a case by
case basis.