Cargo.toml: update dependencies and raise MSRV to 1.65.0
We really had to do some gymnastics to actually be able to
compile threescalers with versions prior to 1.65, so let's
just accept reality.

We are also upgrading reqwest, and using it will significantly
bump the MSRV even if we didn't upgrade, so we're also taking
the opportunity to use the latest 0.12 series.