lib: allow blanket clippy restriction lints group
In recent toolchains clippy whines about warning on the restriction
group and suggests to have a list of lints checked. However I prefer
to lower or disable the lint level with each offending instance, as
it is a practice that helps me avoid go hunting for new lints and
instead just disable or lower their warning level in my editor.

The problem now is that since the code denies all clippy::all lints,
I cannot disable the lint to stop using the restriction group via
the command line, even if I enable the group in the command line
itself. This is an usability wart for me, and the only way to work
around it is by allowing this specific lint in the code. If I was
a clippy developers I'd consider it fine to disable the lint in the
CLI since the cause it triggers is also found in the CLI, but it
may need some extra work to support that and may feel confusing.